refactor(editor-ui): moving JSON view actions into separate components
